+++ types/tools/tool-settings.d.ts
@@ -1,6 +1,6 @@
 import { ToolConfig } from './tool-config';
 import { ToolConstructable, BlockToolData } from './index';
-import { PopoverItemDefaultParams, PopoverItemSeparatorParams, PopoverItemParams } from '../configs';
+import { PopoverItemDefaultParams, PopoverItemSeparatorParams, PopoverItemHtmlParams } from '../configs';
 
 /**
  * Tool may specify its toolbox configuration
@@ -58,9 +58,14 @@ export type TunesMenuConfigDefaultItem = PopoverItemDefaultParams & {
 export type TunesMenuConfigSeparatorItem = PopoverItemSeparatorParams;
 
 /**
+ * Represents single Tunes Menu item with custom HTML contect
+ */
+export type TunesMenuConfigHtmlItem = PopoverItemHtmlParams;
+
+/**
  * Union of all Tunes Menu item types
  */
-export type TunesMenuConfigItem = TunesMenuConfigDefaultItem | TunesMenuConfigSeparatorItem;
+export type TunesMenuConfigItem = TunesMenuConfigDefaultItem | TunesMenuConfigSeparatorItem | TunesMenuConfigHtmlItem;
 
 /**
  * Tool may specify its tunes configuration